Export data/plotting functions, create time and prefix arg
Create new shrimp_funcs.py with data/plotting functions that can be used by
a variety of experimental brine shrimp setups without repeating too much
code. Add a time argument so that the sim time doesn't have to be hard coded.
Add a prefix arg so the filenames of output can be adjusted.