Adds read_comsol_vtu_data, test, many vtu bug fixes, passes test
read_comsol_vtu_data has been added to Planktos to call data_IO and
load comsol vtu data into the environment. A test was created within
test_fluid_read.py to test this method, assuming that data is available
in tests/data/comsol with the name vtu_test_data.txt. The test will be
skipped if the directory DNE and an error will be thrown if the
directory exists but the file does not. Many bugs were discovered in the
data_IO load vtu routine by the testing process; these have been fixed
and the method now passes all tests.