Renames/defines mv_swarm as per #28, get_movement->update_positions
mv_swarm.py is now motion.py, which is envisioned to be a library of
solvers for particle motion of various sorts (RK4, SDE, etc.). This can
then be accessed when solving for the FTLE and the like. Along with
this, get_movement has returned to update_positions which is expected to
set self.positions. This is more inline with the general expectation for
OOP, and will make things simplier if, say, a solver returns the actual
position after dt instead of the change in position.