Bug fixes in FTLE, flexible finite diff, work still needed
While this commit contains some rather important bug fixes and a
paradigm shift toward a flexible finite difference stencil, there is
still a discrepency in the 2D results versus VisIt toward the right hand
side of the channel boundary. The values are really big for some
reason. Initial hypothesis is that somehow the domain boundary is
playing a role, though that does not necessarily seem to be the case
now.