Key bug fix in RK45, much better envir plotting
One number in the RK45 Butcher tableau was mistyped, leading to errors.
The solver is now more robust to continuing after solving to and end
time as well. Plotting in the environment class has now gotten far
better, with correct aspect ratio in all 2D cases and option to add
clipping to the colorbars.