Finally something that works - no numpydoc
autosummary seems irreparably broken with respect to methods of a class
that has been imported to a package. Since numpydoc relies on
autosummary, this means it is unusuable for the time being. Falling back
on napoleon, things now work even if the entire environment and swarm
classes will necessarily need to be on the same page for now.