Fixes bug in calculate_FTLE, props has empty init
Formerly, props was initialized with initial agent positions.
This was both unnecessary (it's in the history) and causes problems
if the size of the swarm should change... e.g., in FTLE
calculations. This commit also fixes a bug in calculate_FTLE
where the grid size is initialized with more or fewer agents than
the original, passed in swarm, but props is only defined for the
former number of agents. Props is now redefined based on the first
agent in the dataframe, since all agents should be equivalent at
the FTLE start time.