3D plotting bug fix, ib_mesh color choice, increment version
3D plotting was completely broken on Windows. Ultimate, the fix was that
creating a 3D axes via mplot3d no longer works - you have to set
projection to 3d when creating a subplot. The environment class now has
an attribute that sets the color of the ib_mesh. This defaults to black
in 2D and darkgray in 3D. It was previously green in 3D, which did not
work well with the new green agents. A minor version number has
incremented to reflect these bug fixes.