Closes #62. Bug fix long _apply_internal_BC recursion
In long recursions of _apply_internal_BC (or _project_and_slide), it is
possible to go beyond the reduced mesh set that was based off of the
distance from the initial point of intersection. This was not previously
apparent because the search distance was rather large - with the new,
tighter search distance, this bug became apparent in ex_ib2d_sticky. It
primarily affects cases where fast-moving particles encounter concave
boundaries.