Adds explanation for dist btwn lines algorithm
The geometry behind finding the closest distance between two lines is
now explained in a markdown file with latex located in docs/notes. For
uniformity, the tex file that used to explain the equations of motion
for various Re has been converted to markdown and added to this
directory as well. Code comments have been added to make things more
clear.