Newton's mthd now LSQ; accuracy fix on mesh elem edge
The edge finding problem was previously misidentified as a root-finding
problem when it is actually a non-linear LSQ problem. The result is that
it was troublesome for Newton's method in many settings since the
distance function does not go negative. The projected point was formerly
being used to specify where the agent reached the edge of the mesh
element, but that is much less accurate than just using the interpolated
edge of the mesh element (since the time of reaching the edge is subject
to numerical error based on tolerance of LSQ).