Major math fix related to sum of velocities
dx/dt should be the sum of velocities along the direction of the mesh
element (due only to agent movement projection) and orthogonal to the
direction of the mesh element (due only to mesh element movement). This
is a bunch of vector calculus that results in a non-homogeneous ODE to
solve for agent position. Thus the proj_to_pt routine has been
completely reworked and an explanation has been included.