Bug fix in project_and_slide_moving w/ nonlinear edge movement
Previously, we were not correctly detecting all movement past the edge
of mesh elements because mesh deformation is not monotonic and neither
is agent movement parallel to the mesh element. This can cause the agent
to leave the mesh and then come back on it during the same time step.
This has hopefully been fixed. Much math has been added to explain the
fix.